Apply a number of code cleanup refactorings
- prefer diamond generics references
- prefer Map.of, Set.of, List.of etc rather than static initializers with anonymous inner classes or unnecessary utils
- prefer expression lambdas where possible

+++ base/src/test/java/com/thoughtworks/go/util/FileSizeUtilsTest.java
@@ -17,34 +17,34 @@ package com.thoughtworks.go.util;
 
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.Matchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
 
 public class FileSizeUtilsTest {
 
     @Test
     public void shouldConvertBytes() {
-        assertThat(FileSizeUtils.byteCountToDisplaySize((long) 1023), is("1023 bytes"));
+        assertThat(FileSizeUtils.byteCountToDisplaySize(1023), is("1023 bytes"));
     }
 
     @Test
     public void shouldConvertBytesToKilo() {
-        assertThat(FileSizeUtils.byteCountToDisplaySize((long) (1024 + 512)), is("1.5 KB"));
+        assertThat(FileSizeUtils.byteCountToDisplaySize(1024 + 512), is("1.5 KB"));
     }
 
     @Test
     public void shouldOnlyKeep() {
-        assertThat(FileSizeUtils.byteCountToDisplaySize((long) (1024 + 512 + 256)), is("1.8 KB"));
+        assertThat(FileSizeUtils.byteCountToDisplaySize(1024 + 512 + 256), is("1.8 KB"));
     }
 
     @Test
     public void shouldConvertBytesToMega() {
-        assertThat(FileSizeUtils.byteCountToDisplaySize((long) (1024 * 1024)), is("1.0 MB"));
+        assertThat(FileSizeUtils.byteCountToDisplaySize(1024 * 1024), is("1.0 MB"));
     }
 
     @Test
     public void shouldConvertBytesToMegaForFloat() {
-        assertThat(FileSizeUtils.byteCountToDisplaySize((long) (1 * 1024 * 1024 + 512 * 1024)), is("1.5 MB"));
+        assertThat(FileSizeUtils.byteCountToDisplaySize(1 * 1024 * 1024 + 512 * 1024), is("1.5 MB"));
     }
 
     @Test
